requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.2.2
pillow==10.0.0
drf-yasg==1.21.7
faker
